Refactor rate limit configuration and improve error handling in customer and user controllers
- Updated rate limit checks to allow negative values and set defaults accordingly in gobe.go.
- Changed import path for fscm in customers_controller.go to use gdbasez bridge.
- Updated APIWrapper type in CustomerController to use ClientModel alias.
- Modified response types in customer controller methods to use ClientModel.
- Changed ExpiresIn and RefreshExpiresIn types in AuthResponse to int64 for better compatibility.
- Simplified user controller methods by consolidating error handling and response formatting.
- Added new ServerController for managing system endpoints including health check, ping, version, and config.
- Introduced new types for responses in manage controller to standardize API responses.
- Updated manage routes to use the new ServerController methods.
- Created gdbasez bridge for cleaner model definitions and service instantiation.
- Commented out sensitive log statements in config loading to enhance security.

+++ internal/app/controllers/sys/federation/users/types.go
@@ -57,8 +57,8 @@ type AuthResponse struct {
 	TokenType        string      `json:"token_type"`
 	AccessToken      string      `json:"access_token"`
 	RefreshToken     string      `json:"refresh_token"`
-	ExpiresIn        int         `json:"expires_in"`
-	RefreshExpiresIn int         `json:"refresh_expires_in"`
+	ExpiresIn        int64       `json:"expires_in"`
+	RefreshExpiresIn int64       `json:"refresh_expires_in"`
 	User             UserSummary `json:"user"`
 }
